requirements.txt
discord.py==2.3.2
asyncpg==0.29.0
groq==0.9.0
google-generativeai==0.7.2
Pillow==10.3.0
python-dotenv==1.0.1
aiohttp==3.9.5
httpx==0.25.2